feat: finished identity service and added auth on auction service

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -48,12 +49,13 @@ namespace AuctionService.Controllers
        Ok(_mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> Create(CreateAuctionDto auctionDto)
     {
       var auction = _mapper.Map<Auction>(auctionDto);
 
-      auction.Seller = "Test";
+      auction.Seller = User.Identity.Name;
 
       _context.Auctions.Add(auction);
 
@@ -69,6 +71,7 @@ namespace AuctionService.Controllers
           new { auction.Id }, _mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -77,6 +80,8 @@ namespace AuctionService.Controllers
         .FirstOrDefaultAsync(x => x.Id == id);
 
       if (auction is null) return NotFound();
+
+      if (auction.Seller != User.Identity?.Name) return Forbid();
 
       auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
       auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -94,6 +99,7 @@ namespace AuctionService.Controllers
 
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -101,6 +107,8 @@ namespace AuctionService.Controllers
         .FindAsync(id);
 
       if (auction is null) return NotFound();
+
+      if (auction.Seller != User.Identity?.Name) return Forbid();
 
       _context.Auctions.Remove(auction);
